MOve IssueListDefDataRegion to issues package
